Use lighter colorPrimary in dark theme
The current colorPrimary, when used in the dark theme, has too low a
contrast to be seen clearly. This is highly evident in the navigation
drawer - when an item is selected, its text cannot be seen at all.

Add a lighter version of colorPrimary to be used in the dark theme.